fuhan

+++ crawler/down_dingdian_fiction.py
@@ -1,12 +1,15 @@
 import requests
 import json
 import re
+import time
 from down_basic import getHtmlByUrl as getHtmlByUrl
 from down_basic import deleteFile as deleteFile
 from down_basic import getBaseDir as getBaseDir
 from down_basic import parse as parse
 from down_basic import isNull as isNull
 from down_basic import saveFile as saveFile
+
+STATUS = 1
 
 BASE_DIR = getBaseDir() + "\\txt\\"
 
@@ -15,8 +18,8 @@ def getAuthor(html):
     author = re.findall(regex_author, html, re.S)
     return author[0]
 
-book_url = "https://www.booktxt.net/1_1894/"
-prefix = 'https://www.booktxt.net/1_1894/'
+book_url = "https://www.booktxt.net/8_8112/"
+prefix = 'https://www.booktxt.net/8_8112/'
 html = getHtmlByUrl(book_url)
 
 book_name_regex = '<h1>(.*?)</h1>'
@@ -27,8 +30,10 @@ print(book_name)
 
 book_txt = BASE_DIR + book_name + '.txt'
 book_catalog_txt = BASE_DIR + book_name + "_catalog.txt"
-deleteFile(book_txt)
-deleteFile(book_catalog_txt)
+# 是否从头开始
+if STATUS == 0:
+    deleteFile(book_txt)
+    deleteFile(book_catalog_txt)
 
 catalog_all_regex = '正文</dt>(.*?)</dl>'
 result = parse(html, catalog_all_regex)
@@ -38,10 +43,19 @@ result = parse(result[0], catalogs_regex)
 isNull(result)
 catalogs = result
 
+catalogs = catalogs[245:]
+
+tmp = 245
 for catalog in catalogs:
-    print(catalog[1] + '...')
+    time.sleep(1.5)
+    print(str(tmp) + ' ' + catalog[1] + '...')
     catalog_url = prefix + catalog[0]
-    html = getHtmlByUrl(catalog_url)
+    for i in range(1,4):
+        try:
+            html = getHtmlByUrl(catalog_url)
+            break
+        except Exception as e:
+            print('连接错误，重新尝试...')
     content_regex = '<div id="content">(.*?)</div>'
     result = parse(html, content_regex)
     isNull(result)
@@ -50,5 +64,6 @@ for catalog in catalogs:
     content = content.replace('&nbsp;', ' ')
     saveFile(book_txt, content, "a", "utf-8")
     saveFile(book_catalog_txt, catalog[1], "a", "utf-8")
+    tmp += 1
 
 print('Over')
